feat(alerts): move create from transaction summary to the header
